reorganize GeneratorMaker
