+++ back_end/src/new_game/generators/generator_maker.py
@@ -1,59 +1,16 @@
 import math
 import os
-from dataclasses import dataclass
 from pathlib import Path
-
-import numpy as np
-import yaml
 
 from src.models.assets import AssetInfo, AssetType
 from src.models.ids import AssetId, BusId, PlayerId
+from src.new_game.util.technology_specs import TechnologySpecs
 from src.tools.random_choice import random_choice
 
 
-@dataclass(frozen=True)
-class TechEvolutionIndicator:
-    base: float
-    change_per_round: float
-    max: float
-    min: float
-
-    def __post_init__(self):
-        assert self.min <= self.base <= self.max, "Base value must be within min and max bounds."
-
-    def value_at_round(self, round_number: int) -> float:
-        """Linear function with clipping at min and max"""
-        val = self.base + self.change_per_round * round_number
-        return float(np.clip(val, self.min, self.max))
-
-
-@dataclass(frozen=True)
-class TechnologySpecs:
-    technology_name: str
-    capacity: TechEvolutionIndicator
-    normalised_power_std: float  # as a fraction of capacity
-    capital_cost_per_mw: TechEvolutionIndicator
-    fixed_cost: TechEvolutionIndicator
-    marginal_cost: TechEvolutionIndicator
-    lifespan: TechEvolutionIndicator
-
-    @classmethod
-    def from_yaml(cls, technology_name: str) -> "TechnologySpecs":
-        with open(f"{os.path.dirname(__file__)}/tech_specs/{technology_name}.yaml") as file:
-            data = yaml.safe_load(file)
-
-        return cls(
-            technology_name=technology_name,
-            capacity=TechEvolutionIndicator(**data["capacity"]),
-            normalised_power_std=data["normalised_power_std"],
-            capital_cost_per_mw=TechEvolutionIndicator(**data["capital_cost_per_mw"]),
-            fixed_cost=TechEvolutionIndicator(**data["fixed_cost"]),
-            marginal_cost=TechEvolutionIndicator(**data["marginal_cost"]),
-            lifespan=TechEvolutionIndicator(**data["lifespan"]),
-        )
-
-
 class GeneratorMaker:
+    path = Path(os.path.dirname(__file__))
+
     @classmethod
     def make_one(cls, asset_id: AssetId, bus_id: BusId, current_round: int, technology_name: str | None = None, player_id: PlayerId = PlayerId.get_npc()) -> AssetInfo:
         """Create a generator with properties based on the current round."""
@@ -92,11 +49,11 @@ class GeneratorMaker:
 
     @classmethod
     def get_available_technologies(cls) -> list[str]:
-        tech_specs_dir = Path(os.path.dirname(__file__)) / "tech_specs"
+        tech_specs_dir = GeneratorMaker.path / "tech_specs"
         yaml_files = tech_specs_dir.rglob("*.yaml")
         return [f.stem for f in yaml_files]
 
     @classmethod
     def _get_technology_spec(cls, technology_name: str) -> TechnologySpecs:
         assert technology_name in cls.get_available_technologies(), f"Technology not available, select from: {cls.get_available_technologies()}."
-        return TechnologySpecs.from_yaml(technology_name)
+        return TechnologySpecs.from_yaml(GeneratorMaker.path, technology_name)
